--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/request/LeaderUpdateRequest.java
@@ -1,12 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.request;
-
-import jakarta.validation.constraints.Min;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class LeaderUpdateRequest {
-    @Min(value = 1)
-    private final int memberId;
-}
