--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/request/ManageParticipationRequest.java
@@ -1,8 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class ManageParticipationRequest {
-    private int memberId;
-}
