--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/request/MeetingParticipationCreateRequest.java
@@ -1,10 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.request;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class MeetingParticipationCreateRequest {
-    private int memberId;
-}
