--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/request/ParticipantReviewCreateRequest.java
@@ -1,9 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class ParticipantReviewCreateRequest {
-    private int reviewerId;
-    private int score;
-}
