--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/response/InterestedMeetingCommandResponse.java
@@ -1,11 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class InterestedMeetingCommandResponse {
-    private int meetingId;
-    private int memberId;
-}
