--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/response/LeaderUpdateResponse.java
@@ -1,10 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class LeaderUpdateResponse {
-    private int meetingId;
-}
