--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/response/ManageParticipationResponse.java
@@ -1,11 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ManageParticipationResponse {
-    private long participationId;
-    private String statusType;
-}
