--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/response/MeetingBalanceCheckResponse.java
@@ -1,11 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class MeetingBalanceCheckResponse {
-    private boolean check;
-    private String message;
-}
