--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/response/MeetingCommandResponse.java
@@ -1,11 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class MeetingCommandResponse {
-    private int meetingId;
-
-}
