--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/response/MeetingParticipationCommandResponse.java
@@ -1,10 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class MeetingParticipationCommandResponse {
-    private long participationId;
-}
