--- linkup/src/main/java/com/learningcrew/linkup/community/command/application/dto/response/ParticipantReviewCommandResponse.java
@@ -1,16 +1,0 @@
-package com.learningcrew.linkup.community.command.application.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-public class ParticipantReviewCommandResponse {
-    private long reviewId;
-    private int revieweeId;
-    private int meetingId;
-    private int score;
-    private LocalDateTime createdAt;
-}
