--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/aggregate/BestPlayerId.java
@@ -1,24 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.aggregate;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Embeddable
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@EqualsAndHashCode
-public class BestPlayerId implements Serializable {
-    @Column(name = "meeting_id")
-    private int meetingId;
-
-    @Column(name = "member_id")
-    private int memberId;
-}
-
