--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/aggregate/InterestedMeetingId.java
@@ -1,20 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.aggregate;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
-import lombok.*;
-
-import java.io.Serializable;
-
-@Embeddable
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-@EqualsAndHashCode
-public class InterestedMeetingId implements Serializable {
-    @Column(name = "meeting_id")
-    private int meetingId;
-    @Column(name = "member_id")
-    private int memberId;
-}
