--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/aggregate/MeetingGender.java
@@ -1,5 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.aggregate;
-
-public enum MeetingGender {
-    M, F, BOTH
-}
