--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/aggregate/MeetingParticipationHistory.java
@@ -1,26 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.aggregate;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "meeting_participation_history")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Builder
-@AllArgsConstructor
-public class MeetingParticipationHistory {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long participationId;
-    private int memberId;
-    @Setter
-    private int meetingId;
-    @Setter
-    private LocalDateTime participatedAt;
-    @Setter
-    private int statusId;
-}
