--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/aggregate/ParticipantReview.java
@@ -1,30 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.aggregate;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "participant_review")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Builder
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class ParticipantReview {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long reviewId;
-    private int reviewerId;
-    private int revieweeId;
-    private int meetingId;
-    private int score;
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    @CreationTimestamp
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-}
