--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/repository/BestPlayerRepository.java
@@ -1,4 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.repository;
-
-public interface BestPlayerRepository {
-}
