--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/repository/MeetingParticipationHistoryRepository.java
@@ -1,21 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.repository;
-
-import com.learningcrew.linkup.meeting.command.domain.aggregate.MeetingParticipationHistory;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface MeetingParticipationHistoryRepository {
-    MeetingParticipationHistory save(MeetingParticipationHistory history);
-
-    void delete(MeetingParticipationHistory history);
-
-    List<MeetingParticipationHistory> findByMeetingIdAndStatusId(int meetingId, int statusId);
-
-    void flush();
-    Optional<MeetingParticipationHistory> findByMeetingIdAndMemberId(int meetingId, int memberId);
-    List<MeetingParticipationHistory> findAllByMeetingIdAndStatusId(int meetingId, int accepted);
-
-    boolean existsByMeetingIdAndMemberId(int meetingId, int memberId);
-
-}
