--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/repository/MeetingRepository.java
@@ -1,13 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.repository;
-
-import com.learningcrew.linkup.meeting.command.domain.aggregate.Meeting;
-
-import java.util.Optional;
-
-public interface MeetingRepository {
-    Meeting save(Meeting meeting);
-
-    Optional<Meeting> findById(int meetingId);
-
-    void flush();
-}
