--- linkup/src/main/java/com/learningcrew/linkup/community/command/domain/repository/ParticipantReviewRepository.java
@@ -1,9 +1,0 @@
-package com.learningcrew.linkup.community.command.domain.repository;
-
-import com.learningcrew.linkup.meeting.command.domain.aggregate.ParticipantReview;
-
-public interface ParticipantReviewRepository {
-    ParticipantReview save(ParticipantReview review);
-    boolean existsByMeetingIdAndReviewerIdAndRevieweeId(int meetingId, int reviewerId, int revieweeId);
-
-}
