--- linkup/src/main/java/com/learningcrew/linkup/community/command/infrastructure/repository/JpaMeetingRepository.java
@@ -1,11 +1,0 @@
-package com.learningcrew.linkup.community.command.infrastructure.repository;
-
-import com.learningcrew.linkup.meeting.command.domain.aggregate.Meeting;
-import com.learningcrew.linkup.meeting.command.domain.repository.MeetingRepository;
-import org.springframework.context.annotation.Primary;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-@Primary
-public interface JpaMeetingRepository extends MeetingRepository, JpaRepository<Meeting, Integer> {
-
-}
