--- linkup/src/main/java/com/learningcrew/linkup/community/command/infrastructure/repository/JpaParticipantReviewRepository.java
@@ -1,13 +1,0 @@
-package com.learningcrew.linkup.community.command.infrastructure.repository;
-
-import com.learningcrew.linkup.meeting.command.domain.aggregate.ParticipantReview;
-import com.learningcrew.linkup.meeting.command.domain.repository.ParticipantReviewRepository;
-import org.springframework.context.annotation.Primary;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-@Primary
-public interface JpaParticipantReviewRepository extends ParticipantReviewRepository, JpaRepository<ParticipantReview, Long> {
-
-    boolean existsByMeetingIdAndReviewerIdAndRevieweeId(int meetingId, int reviewerId, int revieweeId);
-
-}
