+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/application/service/MeetingParticipationCommandService.java
@@ -1,33 +1,20 @@
 package com.learningcrew.linkup.meeting.command.application.service;
 
-import com.learningcrew.linkup.common.query.mapper.StatusMapper;
 import com.learningcrew.linkup.exception.BusinessException;
 import com.learningcrew.linkup.exception.ErrorCode;
 import com.learningcrew.linkup.linker.command.domain.aggregate.User;
-import com.learningcrew.linkup.linker.command.domain.repository.MemberRepository;
 import com.learningcrew.linkup.linker.command.domain.repository.UserRepository;
 import com.learningcrew.linkup.meeting.command.application.dto.request.MeetingParticipationCreateRequest;
 import com.learningcrew.linkup.meeting.command.domain.aggregate.Meeting;
 import com.learningcrew.linkup.meeting.command.domain.aggregate.MeetingParticipationHistory;
 import com.learningcrew.linkup.meeting.command.domain.repository.MeetingParticipationHistoryRepository;
-import com.learningcrew.linkup.meeting.command.infrastructure.repository.JpaMeetingParticipationHistoryRepository;
-import com.learningcrew.linkup.meeting.query.dto.response.MeetingDTO;
-import com.learningcrew.linkup.meeting.query.dto.response.MeetingParticipationDTO;
-import com.learningcrew.linkup.meeting.query.dto.response.MemberDTO;
-import com.learningcrew.linkup.meeting.query.mapper.MeetingParticipationMapper;
-import com.learningcrew.linkup.meeting.query.service.MeetingParticipationQueryService;
-import com.learningcrew.linkup.meeting.query.service.MeetingQueryService;
-import com.learningcrew.linkup.meeting.query.service.StatusQueryService;
-import com.learningcrew.linkup.notification.command.application.helper.MeetingNotificationHelper;
+import com.learningcrew.linkup.meeting.command.domain.repository.MeetingRepository;
 import com.learningcrew.linkup.notification.command.application.helper.NotificationHelper;
 import com.learningcrew.linkup.place.command.domain.aggregate.entity.Place;
-import com.learningcrew.linkup.place.query.service.PlaceQueryService;
-import com.learningcrew.linkup.point.command.application.dto.response.MeetingPaymentResponse;
-import com.learningcrew.linkup.point.command.application.dto.response.PointTransactionResponse;
+import com.learningcrew.linkup.place.command.domain.repository.PlaceRepository;
 import com.learningcrew.linkup.point.command.domain.aggregate.PointTransaction;
 import com.learningcrew.linkup.point.command.domain.repository.PointRepository;
 import lombok.RequiredArgsConstructor;
-import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -35,265 +22,153 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
-@RequiredArgsConstructor // 의존성 주입
+@RequiredArgsConstructor
 public class MeetingParticipationCommandService {
 
-    private final MeetingParticipationHistoryRepository repository;
-    private final MeetingParticipationMapper mapper;
-    private final MeetingQueryService meetingQueryService;
-    private final ModelMapper modelMapper;
-    private final MeetingParticipationQueryService meetingParticipationQueryService;
-    private final StatusQueryService statusQueryService;
-    private final NotificationHelper notificationHelper;
-    private final MeetingNotificationHelper meetingNotificationHelper;
-    private final JpaMeetingParticipationHistoryRepository jpaRepository;
-    private final MemberRepository memberRepository;
+    private final MeetingParticipationHistoryRepository participationRepository;
     private final UserRepository userRepository;
-    private StatusMapper statusMapper;
-    private final PlaceQueryService placeQueryService;
     private final PointRepository pointRepository;
+    private final NotificationHelper notificationHelper;
+    private final PlaceRepository placeRepository;
+    private final MeetingRepository meetingRepository;
 
-    @Transactional(readOnly = true)
-    public MeetingPaymentResponse checkBalance(int meetingId, int userId) {
-        MeetingDTO meeting = meetingQueryService.getMeeting(meetingId);
-        int placeId = meeting.getPlaceId();
+    private static final int STATUS_PENDING = 1;
+    private static final int STATUS_ACCEPTED = 2;
+    private static final int STATUS_REJECTED = 3;
+    private static final int STATUS_DELETED = 4;
 
-        Place place = placeQueryService.getPlaceById(placeId);
-        int rentalCost = place.getRentalCost();
-        int minUser = meeting.getMinUser();
-        int costPerUser = rentalCost / minUser;
-
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
-
-        boolean hasEnoughPoint = user.getPointBalance() >= costPerUser;
-
-        String message = hasEnoughPoint
-                ? "참가 신청이 가능합니다."
-                : "포인트 잔액이 부족합니다. 최소 필요 포인트: " + costPerUser;
-
-        return new MeetingPaymentResponse(message, user.getPointBalance());
-    }
-
-    /* 모임 참가 신청 */
+    /** 1. 참가 신청 */
     @Transactional
     public long createMeetingParticipation(MeetingParticipationCreateRequest request, Meeting meeting) {
-        MeetingParticipationHistory history
-                = modelMapper.map(request, MeetingParticipationHistory.class);
+        int meetingId = meeting.getMeetingId();
+        int memberId = request.getMemberId();
 
-        LocalDateTime now = LocalDateTime.now();
-
-        /* 회원이 모임에 속해 있는지 확인 */
-        List<Integer> participantsIds = jpaRepository.findByMeetingIdAndStatusId(
-                meeting.getMeetingId(), statusQueryService.getStatusId("ACCEPTED")
-        ).stream().map(MeetingParticipationHistory::getMemberId).toList();
-
-        if (participantsIds.contains(request.getMemberId())) {
+        // 중복 참가 확인
+        if (participationRepository.existsByMeetingIdAndMemberId(meetingId, memberId)) {
             throw new BusinessException(ErrorCode.MEETING_ALREADY_JOINED);
         }
 
-        /* 참가 신청 요청자가 개설자이면 ACCEPTED, 아니면 PENDING 처리 */
-        int statusId;
-        if (meeting.getLeaderId() != request.getMemberId()) {
-            statusId = statusQueryService.getStatusId("PENDING");
-        } else {
-            statusId = statusQueryService.getStatusId("ACCEPTED");
-        }
-
-        /* 모임이 참가 가능한 상태인지 확인 -> pending accepted rejected deleted done 중 pending 뿐 */
-        int meetingStatusId = meeting.getStatusId();
-        if (meetingStatusId == statusQueryService.getStatusId("REJECTED")) {
-            throw new BusinessException(ErrorCode.MEETING_PARTICIPATION_LIMIT_EXCEEDED);
-        }
-        if (meetingStatusId == statusQueryService.getStatusId("DELETED") || meetingStatusId == statusQueryService.getStatusId("DONE")) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "참가 신청할 수 없는 모임입니다.");
-        } // 모임 취소 혹은 진행 완료
-
-        /* 시간과도 비교 (모임이 종료되어야 진행 완료 처리되므로, 모임 진행 중에 신청이 가능할 수 있음) */
-        LocalDateTime allowedUntil = meeting.getDate().atTime(meeting.getStartTime());
-        if (allowedUntil.isBefore(now)) {
+        // 모임 상태 확인
+        if (meeting.getStatusId() == STATUS_REJECTED || meeting.getStatusId() == STATUS_DELETED) {
             throw new BusinessException(ErrorCode.BAD_REQUEST, "참가 신청할 수 없는 모임입니다.");
         }
 
-        history.setMeetingId(meeting.getMeetingId());
-        history.setParticipatedAt(now);
-        history.setStatusId(statusId);
-
-        repository.save(history);
-        repository.flush();
-
-        /* 참가 신청 알림 발송 */
-        if(meeting.getLeaderId() != request.getMemberId()){
-        meetingNotificationHelper.sendParticipationRequestNotification(
-                meeting.getLeaderId(),       // 알림 받을 사람 (모임 개설자)
-                meeting.getMeetingTitle()           // 모임 제목 (바인딩될 {meetingTitle})
-        );
-        }
-
-
-        return history.getParticipationId();
-    }
-
-    @Transactional
-    public long acceptParticipation(MeetingDTO meeting, int memberId) {
-        // 1. 참가 내역 조회
-        int meetingId = meeting.getMeetingId();
-        MeetingParticipationHistory participation = jpaRepository.findByMeetingIdAndMemberId(meetingId, memberId);
-
-        if (participation == null || participation.getStatusId() != statusQueryService.getStatusId("PENDING") ) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "승인 가능한 참가 신청 내역이 없습니다.");
-        }
-
-        // 2. 모임이 참가 신청 승인 가능한 상태인지 확인
-        int participantsCount = meetingParticipationQueryService.getParticipantsByMeetingId(meetingId).size();
-
-        // 정원 확인
-        if (participantsCount >= meeting.getMaxUser()) {
-            throw new BusinessException(ErrorCode.MEETING_PARTICIPATION_LIMIT_EXCEEDED);
-        }
-
-        // status 확인
-        if (meeting.getStatusType().equals("모집 완료")) {
-            throw new BusinessException(ErrorCode.MEETING_PARTICIPATION_LIMIT_EXCEEDED);
-        }
-
-        if (meeting.getStatusType().equals("모임 취소") || meeting.getStatusType().equals("모임 진행 완료")) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "참가 신청 내역을 승인할 수 없는 모임입니다.");
-        }
-
-        // 모임 시작 시간 확인
+        // 시간 확인
         if (meeting.getDate().atTime(meeting.getStartTime()).isBefore(LocalDateTime.now())) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "참가 신청 내역을 승인할 수 없는 모임입니다.");
-        }
-        try {
-            payParticipation(meetingId, memberId);
-        } catch (BusinessException e) {
-            throw new BusinessException(ErrorCode.INSUFFICIENT_BALANCE, "포인트 부족으로 참가 승인을 할 수 없습니다.");
+            throw new BusinessException(ErrorCode.BAD_REQUEST, "이미 시작된 모임에는 참가할 수 없습니다.");
         }
 
-        // 3. 참가 승인 처리
-        participation.setStatusId(statusQueryService.getStatusId("ACCEPTED"));
-        MeetingParticipationDTO dto = modelMapper.map(participation, MeetingParticipationDTO.class);
-        dto.setStatusType("ACCEPTED");
+        int statusId = (meeting.getLeaderId() == memberId) ? STATUS_ACCEPTED : STATUS_PENDING;
 
-        repository.save(participation);
-        repository.flush();
+        MeetingParticipationHistory participation = MeetingParticipationHistory.builder()
+                .meetingId(meetingId)
+                .memberId(memberId)
+                .participatedAt(LocalDateTime.now())
+                .statusId(statusId)
+                .build();
 
-
-        /* 참가 승인 알림 발송 */
-        meetingNotificationHelper.sendParticipationAcceptNotification(
-                memberId,       // 알림 받을 사람 (모임 개설자)
-                meeting.getMeetingTitle()           // 모임 제목 (바인딩될 {meetingTitle})
-        );
-
-
+        participationRepository.save(participation);
+        notificationHelper.sendNotification(meeting.getLeaderId(), 1, 1);
 
         return participation.getParticipationId();
     }
 
-    private PointTransactionResponse payParticipation(int meetingId, int memberId) {
-        MeetingDTO meeting = meetingQueryService.getMeeting(meetingId);
-        Integer placeId = meeting.getPlaceId();
-        if(placeId == null) {
-            return new PointTransactionResponse("결제할 장소가 없는 모임입니다.",
-                    userRepository.findById(memberId)
-                            .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND))
-                            .getPointBalance());
-        }
-        Place place = placeQueryService.getPlaceById(placeId);
-        int rentalCost = place.getRentalCost();
-        int minUser = meeting.getMinUser();
-        int amountPerPerson = rentalCost / minUser;
-
-        User user = userRepository.findById(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("유저를 찾을 수 없습니다."));
-
-        if (user.getPointBalance() < amountPerPerson) {
-            throw new BusinessException(ErrorCode.INSUFFICIENT_BALANCE, "포인트가 부족합니다.");
-        }
-
-
-        user.subtractPointBalance(amountPerPerson);
-        userRepository.save(user);
-
-        PointTransaction transaction = new PointTransaction(
-                null,
-                memberId,
-                -amountPerPerson,
-                "PAYMENT",
-                null
-        );
-        pointRepository.save(transaction);
-
-        return new PointTransactionResponse("결제가 완료되었습니다.", user.getPointBalance());
-    }
-
+    /** 2. 참가 승인 */
     @Transactional
-    public long rejectParticipation(MeetingDTO meeting, int memberId) {
-        // 1. 참가 내역 조회
+    public long acceptParticipation(Meeting meeting, int memberId) {
         int meetingId = meeting.getMeetingId();
-        MeetingParticipationHistory participation = jpaRepository.findByMeetingIdAndMemberId(meetingId, memberId);
-//        MeetingParticipationDTO participation = mapper.selectHistoryByMeetingIdAndMemberId(meetingId, memberId);
-        if (participation == null || participation.getStatusId() != statusQueryService.getStatusId("PENDING") ) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "거절 가능한 참가 신청 내역이 없습니다.");
+
+        MeetingParticipationHistory participation = participationRepository
+                .findByMeetingIdAndMemberId(meetingId, memberId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND, "참가 정보를 찾을 수 없습니다."));
+
+        if (participation.getStatusId() != STATUS_PENDING) {
+            throw new BusinessException(ErrorCode.BAD_REQUEST, "승인 가능한 참가 신청이 없습니다.");
         }
 
-        // 2. 모임이 참가 신청 거절 가능한 상태인지 확인
-        if (meeting.getStatusType().equals("모임 취소") || meeting.getStatusType().equals("모임 진행 완료")) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "참가 신청 내역을 거절할 수 없는 모임입니다.");
+        List<MeetingParticipationHistory> acceptedList =
+                participationRepository.findByMeetingIdAndStatusId(meetingId, STATUS_ACCEPTED);
+
+        if (acceptedList.size() >= meeting.getMaxUser()) {
+            throw new BusinessException(ErrorCode.MEETING_PARTICIPATION_LIMIT_EXCEEDED);
         }
 
-        // 모임 시작 시간 확인
-        if (meeting.getDate().atTime(meeting.getStartTime()).isBefore(LocalDateTime.now())) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "참가 신청 내역을 거절할 수 없는 모임입니다");
+        if (meeting.getStatusId() == STATUS_DELETED ||
+                meeting.getDate().atTime(meeting.getStartTime()).isBefore(LocalDateTime.now())) {
+            throw new BusinessException(ErrorCode.BAD_REQUEST, "모임 상태가 유효하지 않아 승인할 수 없습니다.");
         }
 
-        // 3. 참가 거절 처리
-        MeetingParticipationDTO dto = modelMapper.map(participation, MeetingParticipationDTO.class);
-        dto.setStatusType("거절");
+        payParticipation(meeting, memberId);
 
-        participation.setStatusId(statusQueryService.getStatusId("REJECTED"));
-        repository.save(participation);
-        repository.flush();
+        participation.setStatusId(STATUS_ACCEPTED);
+        participationRepository.save(participation);
 
-
-        /* 참가 거절 알림 발송 */
-        meetingNotificationHelper.sendParticipationRejectNotification(
-                memberId,       // 알림 받을 사람 (모임 개설자)
-                meeting.getMeetingTitle()           // 모임 제목 (바인딩될 {meetingTitle})
-        );
-
+        notificationHelper.sendNotification(memberId, 2, 1);
         return participation.getParticipationId();
     }
 
+    /** 3. 참가 거절 */
     @Transactional
-    public long deleteMeetingParticipation(MeetingParticipationDTO history) {
+    public long rejectParticipation(Meeting meeting, int memberId) {
+        int meetingId = meeting.getMeetingId();
+
+        MeetingParticipationHistory participation = participationRepository
+                .findByMeetingIdAndMemberId(meetingId, memberId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.NOT_FOUND, "참가 정보를 찾을 수 없습니다."));
+
+        if (participation.getStatusId() != STATUS_PENDING) {
+            throw new BusinessException(ErrorCode.BAD_REQUEST, "거절 가능한 참가 신청이 없습니다.");
+        }
+
+        if (meeting.getStatusId() == STATUS_DELETED ||
+                meeting.getDate().atTime(meeting.getStartTime()).isBefore(LocalDateTime.now())) {
+            throw new BusinessException(ErrorCode.BAD_REQUEST, "모임 상태가 유효하지 않아 거절할 수 없습니다.");
+        }
+
+        participation.setStatusId(STATUS_REJECTED);
+        participationRepository.save(participation);
+        return participation.getParticipationId();
+    }
+
+    /** 4. 참가 취소 (soft delete) */
+    @Transactional
+    public long deleteMeetingParticipation(MeetingParticipationHistory history) {
         if (history == null) {
-            throw new BusinessException(ErrorCode.NOT_FOUND, "참여 정보가 없습니다.");
+            throw new BusinessException(ErrorCode.NOT_FOUND, "참가 정보를 찾을 수 없습니다.");
         }
 
-        MeetingParticipationHistory entity = modelMapper.map(history, MeetingParticipationHistory.class);
-        entity.setStatusId(statusQueryService.getStatusId("DELETED"));
-        repository.save(entity);
-        repository.flush();
-
-        history.setStatusType("참가 취소"); // soft delete
+        history.setStatusId(STATUS_DELETED);
+        participationRepository.save(history);
         return history.getParticipationId();
+    }
+
+    /** 5. 참가 가능 여부 확인 */
+    @Transactional(readOnly = true)
+    public void validateBalance(int meetingId, int userId) {
+        Meeting meeting = meetingRepository.findById(meetingId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND));
+
+        validateBalance(meeting, userId);
     }
 
     @Transactional(readOnly = true)
-    public void validateBalance(int meetingId, int userId) {
-        MeetingDTO meeting = meetingQueryService.getMeeting(meetingId);
-        Integer placeId = meeting.getPlaceId();
-        if(placeId == null) {
-            return;
+    public void validateBalance(Meeting meeting, int userId) {
+        int costPerUser = 0;
+
+        // 장소가 지정된 경우만 비용 계산
+        if (meeting.getPlaceId() != null) {
+            Place place = placeRepository.findById(meeting.getPlaceId())
+                    .orElseThrow(() -> new BusinessException(ErrorCode.PLACE_NOT_FOUND));
+
+            int minUser = meeting.getMinUser();
+            if (minUser <= 0) {
+                throw new BusinessException(ErrorCode.BAD_REQUEST, "최소 인원이 1명 이상이어야 합니다.");
+            }
+
+            costPerUser = place.getRentalCost() / minUser;
         }
 
-        Place place = placeQueryService.getPlaceById(placeId);
-        int rentalCost = place.getRentalCost();
-        int minUser = meeting.getMinUser();
-        int costPerUser = rentalCost / minUser;
-
+        // 사용자 확인 및 포인트 체크
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
 
@@ -304,6 +179,33 @@ public class MeetingParticipationCommandService {
     }
 
 
+
+    /** 포인트 차감 처리 */
+    private void payParticipation(Meeting meeting, int memberId) {
+        int placeId = meeting.getPlaceId();
+
+        Place place = placeRepository.findById(placeId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.PLACE_NOT_FOUND));
+
+        int rentalCost = place.getRentalCost();
+
+        List<MeetingParticipationHistory> accepted =
+                participationRepository.findByMeetingIdAndStatusId(meeting.getMeetingId(), STATUS_ACCEPTED);
+
+        int numberOfParticipants = accepted.size() + 1;
+        int costPerUser = rentalCost / numberOfParticipants;
+
+        User user = userRepository.findById(memberId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.USER_NOT_FOUND));
+
+        if (user.getPointBalance() < costPerUser) {
+            throw new BusinessException(ErrorCode.INSUFFICIENT_BALANCE);
+        }
+
+        user.subtractPointBalance(costPerUser);
+        userRepository.save(user);
+
+        PointTransaction transaction = new PointTransaction(null, memberId, costPerUser, "PAYMENT", null);
+        pointRepository.save(transaction);
+    }
 }
-
-
