+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/application/service/ParticipantReviewCommandService.java
@@ -3,89 +3,89 @@ package com.learningcrew.linkup.meeting.command.application.service;
 import com.learningcrew.linkup.exception.BusinessException;
 import com.learningcrew.linkup.exception.ErrorCode;
 import com.learningcrew.linkup.meeting.command.application.dto.request.ParticipantReviewCreateRequest;
+import com.learningcrew.linkup.meeting.command.domain.aggregate.Meeting;
+import com.learningcrew.linkup.meeting.command.domain.aggregate.MeetingParticipationHistory;
 import com.learningcrew.linkup.meeting.command.domain.aggregate.ParticipantReview;
+import com.learningcrew.linkup.meeting.command.domain.repository.MeetingParticipationHistoryRepository;
+import com.learningcrew.linkup.meeting.command.domain.repository.MeetingRepository;
 import com.learningcrew.linkup.meeting.command.domain.repository.ParticipantReviewRepository;
-import com.learningcrew.linkup.meeting.query.dto.response.MeetingDTO;
-import com.learningcrew.linkup.meeting.query.dto.response.MeetingParticipationDTO;
-import com.learningcrew.linkup.meeting.query.dto.response.MemberDTO;
-import com.learningcrew.linkup.meeting.query.dto.response.ParticipantReviewDTO;
-import com.learningcrew.linkup.meeting.query.service.MeetingParticipationQueryService;
-import com.learningcrew.linkup.meeting.query.service.MeetingQueryService;
-import com.learningcrew.linkup.meeting.query.service.ParticipantReviewQueryService;
-import com.learningcrew.linkup.meeting.query.service.StatusQueryService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ParticipantReviewCommandService {
-    static final int REVIEW_DUE_DAY = 1;
+
+    private static final int REVIEW_DUE_DAY = 1;
     private static final int MIN_SCORE = 1;
     private static final int MAX_SCORE = 5;
+    private static final int STATUS_DONE = 5;
 
-    private final ParticipantReviewRepository repository;
-    private final MeetingParticipationQueryService participationQueryService;
-    private final MeetingQueryService meetingQueryService;
-    private final ParticipantReviewQueryService participantReviewQueryService;
-    private final StatusQueryService statusQueryService;
+    private final ParticipantReviewRepository reviewRepository;
+    private final MeetingRepository meetingRepository;
+    private final MeetingParticipationHistoryRepository participationRepository;
 
     @Transactional
     public long createParticipantReview(ParticipantReviewCreateRequest request, int revieweeId, int reviewerId, int meetingId) {
-        List<Integer> participants = participationQueryService.getHistories(meetingId, statusQueryService.getStatusId("DONE") )
-                .stream()
-                .map(MeetingParticipationDTO::getMemberId)
+
+        // 1. 모임 존재 여부
+        Meeting meeting = meetingRepository.findById(meetingId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.MEETING_NOT_FOUND));
+
+        // 2. 참가 완료된 사람들만 리뷰 가능 (STATUS_DONE = 5)
+        List<MeetingParticipationHistory> doneParticipants =
+                participationRepository.findByMeetingIdAndStatusId(meetingId, STATUS_DONE);
+
+        List<Integer> participantIds = doneParticipants.stream()
+                .map(MeetingParticipationHistory::getMemberId)
                 .toList();
 
-        if (!participants.contains(reviewerId) || !participants.contains(revieweeId)) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "유효하지 않은 참가자 평가 요청입니다.");
+        if (!participantIds.contains(reviewerId) || !participantIds.contains(revieweeId)) {
+            throw new BusinessException(ErrorCode.BAD_REQUEST, "리뷰 대상이 모임에 참여하지 않았습니다.");
         }
 
-        // 완료된 모임일때만 리뷰 작성 가능
-        MeetingDTO meetingDTO = meetingQueryService.getMeeting(meetingId);
-        if (!meetingDTO.getStatusType().equals("모임 진행 완료")) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "모임원 평가를 작성할 수 없는 모임입니다.");
+        // 3. 모임이 종료 상태인지 확인
+        if (meeting.getStatusId() != STATUS_DONE) {
+            throw new BusinessException(ErrorCode.REVIEW_NOT_ALLOWED, "모임이 종료되지 않았습니다.");
         }
 
-        // 평가는 모임 종료 후 1일 이내에 작성해야 함
-        LocalDateTime reviewDue = meetingDTO.getDate().plusDays(REVIEW_DUE_DAY).atTime(meetingDTO.getEndTime());
-        if (reviewDue.isBefore(LocalDateTime.now())) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "모임원 평가 기간이 지났습니다.");
+        // 4. 작성 기한 확인
+        LocalDateTime deadline = meeting.getDate().plusDays(REVIEW_DUE_DAY).atTime(meeting.getEndTime());
+        if (deadline.isBefore(LocalDateTime.now())) {
+            throw new BusinessException(ErrorCode.REVIEW_ALREADY_EXISTS, "평가 가능 기간이 지났습니다.");
         }
 
-        // 자기 자신 평가 불가
+        // 5. 자기 자신은 평가 불가
         if (reviewerId == revieweeId) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "유효하지 않은 참가자 평가 요청입니다.");
+            throw new BusinessException(ErrorCode.BAD_REQUEST, "자기 자신은 평가할 수 없습니다.");
         }
 
-        // 평가 점수 유효성 검사
-        if (request.getScore() < MIN_SCORE || request.getScore() > MAX_SCORE) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "유효하지 않은 평가 점수입니다.");
+        // 6. 중복 리뷰 확인
+        boolean isAlreadyReviewed = reviewRepository.existsByMeetingIdAndReviewerIdAndRevieweeId(
+                meetingId, reviewerId, revieweeId);
+        if (isAlreadyReviewed) {
+            throw new BusinessException(ErrorCode.REVIEW_ALREADY_EXISTS);
         }
 
-        // 중복 리뷰 체크 로직
-        ParticipantReviewDTO reviewDTO = participantReviewQueryService.getReview(meetingId, reviewerId, revieweeId);
-        if (reviewDTO != null) {
-            throw new BusinessException(ErrorCode.BAD_REQUEST, "이미 리뷰 등록된 사용자입니다.");
+        // 7. 점수 유효성
+        int score = request.getScore();
+        if (score < MIN_SCORE || score > MAX_SCORE) {
+            throw new BusinessException(ErrorCode.INVALID_REVIEW_SCORE);
         }
 
-
-        // 리뷰 등록
+        // 8. 저장
         ParticipantReview review = ParticipantReview.builder()
+                .meetingId(meetingId)
                 .reviewerId(reviewerId)
                 .revieweeId(revieweeId)
-                .meetingId(meetingId)
-                .score(request.getScore())
+                .score(score)
                 .createdAt(LocalDateTime.now())
                 .build();
 
-        return repository.save(review).getReviewId();
+        return reviewRepository.save(review).getReviewId();
     }
-
-
 }
