+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/domain/repository/MeetingParticipationHistoryRepository.java
@@ -1,10 +1,9 @@
 package com.learningcrew.linkup.meeting.command.domain.repository;
 
-import com.learningcrew.linkup.linker.command.domain.aggregate.Member;
 import com.learningcrew.linkup.meeting.command.domain.aggregate.MeetingParticipationHistory;
-import com.learningcrew.linkup.meeting.query.dto.response.MeetingParticipationDTO;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface MeetingParticipationHistoryRepository {
     MeetingParticipationHistory save(MeetingParticipationHistory history);
@@ -14,8 +13,9 @@ public interface MeetingParticipationHistoryRepository {
     List<MeetingParticipationHistory> findByMeetingIdAndStatusId(int meetingId, int statusId);
 
     void flush();
-
-    MeetingParticipationHistory findByMeetingIdAndMemberId(int meetingId, int memberId);
-
+    Optional<MeetingParticipationHistory> findByMeetingIdAndMemberId(int meetingId, int memberId);
     List<MeetingParticipationHistory> findAllByMeetingIdAndStatusId(int meetingId, int accepted);
+
+    boolean existsByMeetingIdAndMemberId(int meetingId, int memberId);
+
 }
