+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/infrastructure/repository/JpaInterestedMeetingRepository.java
@@ -7,4 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 @Primary
 public interface JpaInterestedMeetingRepository extends InterestedMeetingRepository, JpaRepository<InterestedMeeting, Integer> {
+    boolean existsByInterestedMeetingId_MeetingIdAndInterestedMeetingId_MemberId(int meetingId, int memberId);
 }
