+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/infrastructure/repository/JpaMeetingParticipationHistoryRepository.java
@@ -5,10 +5,17 @@ import com.learningcrew.linkup.meeting.command.domain.repository.MeetingParticip
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Primary
 public interface JpaMeetingParticipationHistoryRepository extends MeetingParticipationHistoryRepository, JpaRepository<MeetingParticipationHistory, Long> {
-    Optional<MeetingParticipationHistory> findByMemberIdAndMeetingIdAndStatusId(int memberId, int meetingId, int statusId);
+
+    List<MeetingParticipationHistory> findByMeetingIdAndStatusId(int meetingId, int statusId);
+
+    Optional<MeetingParticipationHistory> findByMeetingIdAndMemberId(int meetingId, int memberId);
+
+    boolean existsByMeetingIdAndMemberId(int meetingId, int memberId);
+    List<MeetingParticipationHistory> findByMemberIdAndMeetingIdAndStatusId(int memberId, int meetingId, int statusId);
 
 }
