+++ linkup/src/main/java/com/learningcrew/linkup/meeting/command/infrastructure/repository/JpaParticipantReviewRepository.java
@@ -5,8 +5,9 @@ import com.learningcrew.linkup.meeting.command.domain.repository.ParticipantRevi
 import org.springframework.context.annotation.Primary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
 @Primary
 public interface JpaParticipantReviewRepository extends ParticipantReviewRepository, JpaRepository<ParticipantReview, Long> {
+
+    boolean existsByMeetingIdAndReviewerIdAndRevieweeId(int meetingId, int reviewerId, int revieweeId);
 
 }
